Add user_is_instructor - PMT #113636
This extends our auth model beyond just the automatically provided
is_staff concept to be customizable. Users in this system are either an
"instructor", or not. At the moment, an instructor is either Thomas, or
any of our staff.

I've added new is_instructor JS variables, but leaving the old is_staff
variable for now until I change the JS code as well.