Add line transformation model - PMT #114441
This is a more robust way of saving position in JSXGraph than my
cobbled together method of offsets and slopes. This allows the
rotation transformation to work correctly, which is needed for
some things in EconPractice - see referenced PMT.

I'm hoping to embed the transformations within the Graph Viewset,
allowing them to be fetched and updated at that endpoint from the
client.