API: Add nested writability
In the end, the problem I was seeing was because django-rest-framework
uses
[MultiPartRenderer](http://www.django-rest-framework.org/api-guide/renderers/#multipartrenderer)
in testing mode, and this format doesn't support nested data - it
doesn't encode correctly. Because of that, I'm able to just do
`format=json` in the test requests.

The nested serializer code is complicated, but tested, obviously.
I took a look at
[https://github.com/beda-software/drf-writable-nested](drf-writable-nested),
and I couldn't get it to work, but I didn't take too long here since my
custom method works. Still, if this grows too much larger, I may take
another look at the drf-writable-nested library.