Add assessment model - PMT #115086
I'm adding the assessment model here, to be handled through Django's
admin site. This isn't connected yet to the react app, but that will be
a pretty simple step as a read-only resource through
django-rest-framework.

We discussed the possibility of Google Spreadsheet for this
functionality.. that's still on the table and this work doesn't prevent
that, but I think doing things that way may introduce some complexity.
Anyways, for now I'm just getting a straightforward way to create
assessment rules so we can start testing things out from end to end.