Remove Submission.choice
The new assessment system will create a submission by summing up all the
scores of the matching rules. It no longer makes sense to have a single
"choice" field for the user's choice on the submission object, since I'm
not planning on boiling down all the possible chosen states to a single
attribute.