Clean up graph ordering
Order the graphs with respect to the cohort. Update the API tests to
include the topic on create.

I've also removed the override code taken from here:
https://github.com/bfirsh/django-ordered-model/issues/85

Because I'm not sure what issue this was solving. I've removed a test
for the bottom() function, because this isn't actually used in the
custom graph ordering UI.

Tom is seeing graph ordering issues with featured graphs. I can't
reproduce the problem, but I'll see if this fixes it.