Fix JS build command
`npm run build` should run webpack with the prod configuration. This
should fix an error I ran into with the automated build in jenkins:

```
Failed to compile.

configuration
The 'mode' option has not been set, webpack will fallback to
'production' for this value.
Set 'mode' option to 'development' or 'production' to enable defaults
for each environment.
    You can also set it to 'none' to disable any default behavior. Learn
    more: https://webpack.js.org/configuration/mode/

```