Update main.css
Generated by `make scss`. This file is refreshed automatically by jenkins.

This file could possibly be added to .gitignore, but then just running
the project would require npm/scss setup, so I think things are okay
as is. If anything, I would like to get rid of scss, or minimize it's
use, but that's just me.