Reset instructor form to default font size
I think we should use bootstrap's default font size here, as this will
defer to the browser's default font size, which is customizable by the
user. This is reasonably accessible as is, and changing the default font
size for all these elements requires the instructor to scroll around
more to see everything.

https://getbootstrap.com/docs/4.6/content/typography/

On this front, I think a more important area to focus on is enhancing
the accessibility of the SVG graphs that all the students see, as this
is really where we want to focus our accessibility efforts.