Migrate to psycopg 3.x
Django 4.2 adds support for psycopg version 3:
https://docs.djangoproject.com/en/4.2/releases/4.2/#psycopg-3-support

It looks like the recommended way to install this package is via the
binary package:
https://www.psycopg.org/psycopg3/docs/basic/install.html

This method has its advantages and disadvantages. Sometimes it's more
reliable to not use a binary package and build from source, but we'll
see if we run into any issues here and we can adjust as needed.