Add fixed bg color to jsxgraph display
This alleviates issue #2271.

It's not solved completely, but I think the SVG display area should have
a hard-coded color as a background. We are hard-coding all the line
colors themselves, so it's really designed with a white background in
mind.

Additionally, this CSS file was being included redundantly in many
templates: it only needs to be included in base.html if the template
inherits from base.