Move CUIT privacy notice to login page
This really doesn't need to display on every page, and just causes an
annoyance with the page refreshes in new assignment form. Even when you
have the pop-up disabled, it will display briefly as the page loads,
before the JS has a chance to hide it.

So, just keep this on the login page for now.

And, as far as the public student graph view, we aren't using any
cookie or login data here so it's really not relevant.