Refactor Graph.js
The Graph.js file contains the base Graph class, and all the various
graph classes. I'd like to organize this better into each graph class
having its own file, as this will be growing as we add more in the fall.

These changes begin the process of that refactor, just breaking out the
ADASGraph into its own file, and setting things up for the rest of the
changes to happen.