Implement custom graph viewer for assignment steps
This React view had to be customized a bit, because I'm not planning on
directly using the existing Assessment.js form stuff. I want just a more
minimal react view here, so we can handle all the submission stuff in
Django. Also, we don't need to differentiate between instructor/student
in this view, so a lot can be simplified here.

I've added some basic 'actions' events on line move, which create hidden
input fields which can be used for evaluation on submit. Eventually this
can be expanded to all fields.