Fix graph overlap in `lg` breakpoint
This is a quick fix for #2271.

Negative margins are not good practice at all, but the
[CSS breakpoint](https://getbootstrap.com/docs/5.3/layout/breakpoints/)
when this occurs is not typically active.

Going forward, a more robust solution could be to allow the jsxgraph
view itself to have an adaptable width, but that is a big change, as a
lot of the Graph code takes advantage of the fact that we know the exact
graphical view width here of 540px.