Graph.postMake: simplify line event logic
We don't need to seperate the logic between instructor and student here,
as it's basically the same thing happening in both cases, the student's
case just adds a few more custom line events.

In fact, now we want all events to come through in both cases, as this
will be useful for an instructor building an assignment.