Enable assignment tests
`./manage.py test` wasn't including the assignment tests, because
the `__init__.py` file was missing from the directory, so Django's
DiscoverRunner couldn't find the tests automatically.

So, I've added that file so the tests are now run, and fixed the tests
themselves with some of the recent re-factors I've done.