Migrate to better-react-mathjax
There are many react/mathjax component libraries and none of them are
really "officially" maintained, so it's a bit of trial and error to find
one that works best.

I actually wasn't able to directly reproduce the issue in the ECON-660
ticket in my browsers (Firefox / Brave), but I have definitely been
seeing mathjax warnings in the JS console, and just general performance
weirdness with loading these componenents, so I've migrated to a
different library to see if things behave better.