Allow grid options to be optional in JXGBoard
These grid options aren't technically required for displaying the graph,
and aren't marked is `isRequired` in the props, so these changes fix the
case where the code would fail if they aren't present.

In practice, this is not likely to happen as we have populated all grid
types as 0 by default. Still, for most of the graph types that aren't
using the grid yet, it is possible for them to render a JXGBoard without
passing in these options.