Put graph settings checkboxes in two columns
I am just consolidating some vertical space here - it has always
bothered me that the graph-specific settings are pushed down in the page
in the right column of the graph editor. I think there is some empty
space we can re-claim in these general graph settings, possibly with the
dropdown choices as well.