Migrate cobb-douglas hardcoded values to newer "a" variables
Issue #3368

Moving forward, it seems all graphs contain a set of fields which all make
use of the following properties:
* Value
* Name (optional)
* Min/Max (optional)
* Alternate value (optional)

We can encapsulate this into an abstraction called a graph "Field".
It's still to be determined what this will actually look like on the
back-end, I am just trying to understand things conceptually.