Cost Functions: Total: Fix initial values and graph scale
My plan is to actually combine the two cost functions graph types into a
single one with radio buttons. Working up to that, here are a bunch of
fixes for this graph type, correcting the default scale, and also a bug
in calculateBoundingBox().

I've generalized some of the defaults logic into a setDefault() function
we can use for any graphs that have multiple function types or modes.