Simplify beforeunload handler
I've simplified the logic here to listen to any form `change` event. We
can adjust this as needed but that should capture what we want here.

All we need to do to trigger this alert is call `preventDefault()` on the
handler, according to the docs:
https://developer.mozilla.org/en-US/docs/Web/API/Window/beforeunload_event

Also, when querying DOM nodes we have to first wait for the
DOMContentLoaded event to ensure everything is loaded before our code
runs.