Introduce Cohort.students
EconPractice's "Cohort" course model needs a student roster now, so I'm
adding this to the model and will build out the views and tests soon.

I've considered using django_courseaffils for this in the same way we're
doing for Mediathread, but decided it's simpler to just do our own thing
here. django_courseaffils uses the Group model which is not necessary,
and a lot of the functionality there is just outdated or not necessary.
In EconPractice I am fine with our manual course management for now.